+++ R/gtrends.R
@@ -1,16 +1,3 @@
-##-----------------------------------------------------------------------------
-##  Authors:        Philippe Massicotte and Dirk Eddelbuettel
-##  Date modified:  20-03-2014
-##  Description:    TODO
-##-----------------------------------------------------------------------------
-
-## TODO:
-# - better authentication success checks
-# - better query result checks
-# - adding category code (as you mentioned)
-# - restoring ability to plot regions (ahem, I broke that...)
-# - plot googleVis in PDF. Hint: cat(unlist(G1$html), file="tmp.html") -----> print to PDF using system(...)
-
 #' Connect to Google account
 #'
 #' The resulting connection object is also stored in the package-local
@@ -161,8 +148,9 @@ gconnect <- function(usr = NULL, psw = NULL, verbose = FALSE) {
 #'   Multiple keywords are possible using \code{gtrends(c("NHL", "NBA", "MLB", 
 #'   "MLS"))}.
 #'   
-#' @param geo A character variable denoting a geographic region for the query, 
-#'   default to \dQuote{all} for global queries.
+#' @param geo A character vector denoting geographic regions for the query,
+#'   default to \dQuote{all} for global queries. Multiple regions are possible
+#'   using \code{gtrends("NHL", c("CA", "US"))}.
 #'   
 #' @param cat A character denoting the category, defaults to \dQuote{0}.
 #'   
@@ -171,7 +159,7 @@ gconnect <- function(usr = NULL, psw = NULL, verbose = FALSE) {
 #' @param res Resolution of the trending data to be returned. Either \code{week}
 #'   for weekly data or \code{day} for daily data.
 #'   
-#' @param start_date Starting date using yyyy-mm-dd format. Must be breater than
+#' @param start_date Starting date using yyyy-mm-dd format. Must be greater than
 #'   2004-01-01.
 #'   
 #' @param end_date Starting date using yyyy-mm-dd format. Must be before than 
@@ -187,6 +175,8 @@ gconnect <- function(usr = NULL, psw = NULL, verbose = FALSE) {
 #' \dontrun{
 #' ch <- gconnect("usr@gmail.com", "psw")
 #' sport_trend <- gtrends(c("NHL", "NBA", "MLB", "MLS"))
+#' 
+#' sport_trend <- gtrends("NHL", geo = c("CA", "US"))
 #' }
 #' @export
 gtrends <- function(query, geo, cat, ch, ...) {
@@ -195,6 +185,7 @@ gtrends <- function(query, geo, cat, ch, ...) {
     
 }
 
+#' @importFrom utils data
 #' @rdname gtrends
 #' @export
 gtrends.default <- function(query, 
@@ -214,10 +205,23 @@ gtrends.default <- function(query,
             is.vector(query),
             all(res %in% c("week", "day")),
             length(res) == 1,
-            length(query) <= 5)
+            length(query) <= 5, 
+            length(geo) <= 5)
+  
+  if(length(query) > 1 & length(geo) > 1){
+    stop("Can not specify multiple keywords and geo at the same time.", 
+         call. = FALSE)
+  }
+  
+  cmpt <- ifelse(length(query) > 1, "q", "geo")
   
   if(is.null(ch)) stop("You are not signed in. Please log in using gconnect().",
                        call. = FALSE)
+  
+  if (inherits(ch, "CURLHandle") != TRUE) {
+    stop("'ch' arguments has to be result from 'gconnect()'.", 
+         call. = FALSE)
+  }
   
   ## Verify the dates
   start_date <- as.Date(start_date, "%Y-%m-%d")  
@@ -248,24 +252,25 @@ gtrends.default <- function(query,
   query <- paste(query, collapse = ",")
   
   ## Change encoding to utf-8
-  query <- iconv(query, "latin1", "utf-8", sub = "byte")
-  
-  if (inherits(ch, "CURLHandle") != TRUE) {
-    stop("'ch' arguments has to be result from 'gconnect()'.", 
-         call. = FALSE)
+  if (!(Encoding(query) == "UTF-8")) {
+    query <- iconv(query, "latin1", "utf-8", sub = "byte")
   }
   
   data(countries, envir = environment())
   
   countries[, 1] <- as.character(countries[, 1])
   countries[, 2] <- as.character(countries[, 2])
-  countries[which(countries[, "COUNTRY"] == "Namibia"), "CODE"] <- "NA"
+  countries[which(countries[, "country"] == "Namibia"), "code"] <- "NA"
   
-  if (geo != "" && !geo %in% countries[, "CODE"]) {
+  if (geo != "" && !all(geo %in% countries[, "code"]) && !all(geo %in% countries[, "subcode"])) {
     stop("Country code not valid. Please use 'data(countries)' to retreive valid codes.",
          call. = FALSE)
   }
   
+  # https://www.google.com/trends/trendsReport?&q=nhl&geo=US%2C%20BR&cmpt=geo&content=1&export=1
+  
+  geo <- paste(geo, sep = "", collapse = ", ")
+  #geo <- URLencode(geo, reserved = TRUE)
   
   authenticatePage2 <- getURL("http://www.google.com", curl = ch)
   
@@ -275,7 +280,7 @@ gtrends.default <- function(query,
   res <- paste(nmonth, "m", sep = "")
   
   pp <- list(q = query, 
-             cmpt = "q",
+             cmpt = cmpt, 
              content = 1, 
              export = 1,
              date = paste(format(start_date, "%m/%Y"), res),
@@ -329,41 +334,22 @@ summary.gtrends <- function(object, ...) {
 #' @param x A \code{\link{gtrends}} object
 #' @param type A character variable selecting the type of plot;
 #' permissible values are \sQuote{trends} (which is also the
-#' default), \sQuote{regions} and \sQuote{cities}.
-#' @param region A character variable with default
-#' \sQuote{world}. Oher permissible value are country codes like
-#' \sQuote{CA} or \sQuote{GB}, a US Metro code such as \sQuote{US-IL}
-#' or a three-digit code for a continent or sub-continent; see the
-#' help for \link[googleVis]{gvisGeoChart} for details.
-#' @param resolution A character variable selecting the granularity
-#' of the plot; permissble values are \sQuote{countries},
-#' \sQuote{provinces} or \sQuote{metros}.
-#' @param displaymode A character variable indicating the mode of
-#' display, with values \sQuote{auto}, \sQuote{regions} or
-#' \sQuote{markers} with latter preferable for cities.
+#' default), \sQuote{geo}.
+#' @param which Block number containing the geographical data to plot.
 #' @param ind A integer selecting the result set in case of multiple
 #' search terms.
 #' @import googleVis
-#' @import RColorBrewer
 #' @import ggplot2
+#' @importFrom graphics plot
+#' @importFrom stats reshape
+#' @importFrom utils data
 #' @examples 
 #' data("sport_trend")
 #' plot(sport_trend)
 #' @export
-plot.gtrends <- function(x,
-                         type = c("trend", "regions", "topmetros", "cities"),
-                         region = "world",
-                         resolution = c("countries", "provinces", "metros"),
-                         displaymode = c("auto", "regions", "markers"),
-                         ind = 1L,
-                         ...) {
+plot.gtrends <- function(x, type = c("trend", "geo"), which = 5, ind = 1L, ...){
+  
   type <- match.arg(type)
-  
-  resolution <- match.arg(resolution)
-  
-  gvisopt <- list(region = region,
-                  displayMode = "markers",
-                  resolution = resolution)
   
   if (type == "trend") {
     
@@ -388,39 +374,40 @@ plot.gtrends <- function(x,
     
     print(p)
     
-  } else if (type == "regions") {
+  } else if (type == "geo") {
     
-    x <- x[["regions"]][[ind]]
+    stopifnot(ind <= length(x[which]),
+              which >= 5,
+              which <= length(x))
     
-    if(all(is.na(x))) stop("Not enough search volume to show results.", 
+    block <- x[which][[ind]]
+    
+    # Try to find if the requested block contains geographic information.
+    data(locations, envir = environment())
+    
+    if(!any(tolower(block[1, ]) %in% tolower(locations$Name))){
+      
+      message("The requested block does not seems to contain geographical information. Please choose another block.")
+      
+      print(paste(1:length(x), ":", " ", names(x), sep = ""))
+      
+      stop(call. = FALSE)
+      
+    }
+    
+    if(all(is.na(block))) stop("Not enough search volume to show results.", 
                       call. = FALSE)
     
-    df <- data.frame(loc = x[, 1], hits = x[, 2])
+    df <- data.frame(loc = block[, 1], hits = block[, 2])
     
-    plot(gvisGeoChart(df, 'loc', 'hits', options = gvisopt))
-    
-  } else if (type == "topmetros") {
-    
-    x <- x[["topmetros"]][[ind]]
-    
-    if(all(is.na(x))) stop("Not enough search volume to show results.", 
-                      call. = FALSE)
-    
-    df <- data.frame(loc = x[, 1], hits = x[, 2])
-    
-    plot(gvisGeoChart(df, 'loc', 'hits', options = gvisopt))
-    
-  } else if (type == "cities") {
-    
-    x <- x[["cities"]][[ind]]
-    
-    if(all(is.na(x))) stop("Not enough search volume to show results.", 
-                      call. = FALSE)
-    
-    df <- data.frame(loc = x[, 1], hits = x[, 2])
-    
-    plot(gvisGeoChart(df, 'loc', 'hits', options = gvisopt))
-  }
+    plot(gvisGeoChart(df, 
+                      "loc",
+                      "hits",
+                      options = list(region = "world",
+                                     displayMode = "markers",
+                                     resolution = "countries")))
+  } 
+  
   
   invisible(NULL)
 }
@@ -432,12 +419,16 @@ as.zoo.gtrends <- function(x, ...) {
   z
 }
 
+#' @importFrom utils read.csv
+#' @importFrom stats na.omit
 .processResults <- function(resultsText, queryparams) {
   
   #get back to latin1 encoding
   queryparams[1] <- iconv(queryparams[1], "utf-8", "latin1", sub = "byte")
   
   vec <- strsplit(resultsText, "\\\n{2,}")[[1]]
+  
+  headers <- unname(sapply(vec, function(v) strsplit(v, "\\\n")[[1]][1]))
   
   ## Make sure there are some results have been returned.
   if (length(vec) < 2) {
@@ -460,6 +451,8 @@ as.zoo.gtrends <- function(x, ...) {
   weeks <- data.frame(do.call(rbind, strsplit(trend[, 1], " - ")))
   
   trend <- trend[, mapply(is.numeric, trend), drop = FALSE]
+  
+  names(trend) <- unlist(strsplit(queryparams[1], ","), use.names = FALSE)
   
   is_weekly <- all(do.call(c, 
                            lapply(weeks, 
@@ -492,127 +485,57 @@ as.zoo.gtrends <- function(x, ...) {
   
   trend <- cbind(weeks, trend)
   
-  # Verify that all requested keywords have been recevied. Sometimes not
-  # enough data for some requested kw. If that happen, Google will not 
-  # return data.
-  wanted_kw <- make.names(tolower(unlist(strsplit(queryparams[1], ","))))
-  received_kw <- names(trend)[mapply(is.numeric, trend)]
-  
-  if(length(setdiff(wanted_kw, received_kw) != 0)){
-    message(paste("Not enough data for ", setdiff(wanted_kw, received_kw), 
-                  " keyword(s).", sep = "'"))
-  }
-  
+  trend <- na.omit(trend)
   
   #---------------------------------------------------------------------
   # Section to deal with geographical data
   #---------------------------------------------------------------------
   
-  ## results headers -- for 'geo="US"' and three terms, we get 17 results (!!)
-  headers <- unname(sapply(vec, function(v) strsplit(v, "\\\n")[[1]][1]))
-
-  #hit_sum <- colSums(trend[, mapply(is.numeric, trend)])
-  
-  nkw <- length(received_kw)
-  ## first set of blocks: top regions
-  
+  ## block 3+: geographical info
   start <- 3 # Always start at index 3
   
-  start <- seq(start, start + (nkw - 1))
-  
-  reglist <-
-    lapply(start, function(i)
-      read.csv(
-        textConnection(strsplit(vec[i], "\\\n")[[1]]),
-        skip = 1,
-        stringsAsFactors = FALSE
-      ))
-  
-  ## next (optional, if geo==US) block
-  if (queryparams["geo"] == "US") {
-    
-    start <- max(start) + 1
-    start <- seq(start, start + (nkw - 1))
-    
-    #metidx <- grep("Top metros", headers)
-    metlist <-
-      lapply(start, function(i)
-        read.csv(
-          textConnection(strsplit(vec[i], "\\\n")[[1]]),
-          skip = 1,
-          stringsAsFactors = FALSE
-        ))
-  }else{
-    metlist <- NULL
-  }
-  
-  ## next block: top cities
-  start <- max(start) + 1
-  start <- seq(start, start + (nkw - 1))
-  
-  citlist <-
-    lapply(start, function(i)
-      read.csv(
-        textConnection(strsplit(vec[i], "\\\n")[[1]]),
-        skip = 1,
-        stringsAsFactors = FALSE
-      ))
-  
-  ## next block: top searches
-  start <- max(start) + 1
-  start <- seq(start, start + (nkw - 1))
-  
-  schlist <-
-    lapply(start, function(i)
-      read.csv(
-        textConnection(strsplit(vec[i], "\\\n")[[1]]),
-        skip = 1,
-        stringsAsFactors = FALSE,
-        header = FALSE
-      ))
-  
-  ## Set columns names
-  if(length(start) != 0){
-    schlist <- lapply(1:length(start), function(i) {
-      names(schlist[[i]]) = c(headers[start][i], "Hits")
-      schlist[[i]]
-    })
-  }
-  
-  ## nex block: rising searches
-  start <- max(start) + 1
-  start <- seq(start, start + (nkw - 1))
-  
-  rislist <- lapply(start, function(i) {
-    ## broken by design: not a csv when a field can be "+1,900%" with a comma as
-    ## a decimal separator -- so subst out the first comma into a semicolon
-    tt <- sub(",", ";", strsplit(vec[i], "\\\n")[[1]])
-    rising <- read.csv(
-      textConnection(tt),
-      sep = ";",
+  blocks <- lapply(start:length(vec), function(i)
+    read.csv(
+      textConnection(strsplit(vec[i], "\\\n")[[1]]),
       skip = 1,
-      header = FALSE,
-      col.names = c("term", "change"),
       stringsAsFactors = FALSE
-    )
-    rising
-  })
+    ))
   
-  res <- list(
-    query = queryparams,
-    meta = meta,
-    trend = trend,
-    regions = reglist,
-    topmetros = metlist,
-    cities = citlist,
-    searches = schlist,
-    rising = rislist,
-    headers = headers
-  )
+  
+  blocks <- Map(assign, 
+                make.names(headers[start:length(headers)]), 
+                value = blocks)
+  
+  res <- list(query = queryparams,
+              meta = meta,
+              trend = trend,
+              headers = headers)
+
+  
+  res <- append(res, blocks)
+                
+ 
+  # res <- list(
+  #   
+  #   query = queryparams,
+  #   meta = meta,
+  #   
+  #   trend = trend,
+  #   
+  #   regions = res[which(types == "State" | types == "Province")],
+  #   topmetros = res[which(types == "DMA Region")],
+  #   cities = res[which(types == "City")],
+  #   
+  #   # searches = schlist,
+  #   # rising = rislist,
+  #   
+  #   headers = headers
+  # )
   
   # if data was returned monthly, it will not be possible to plot maps
-  res[lapply(res, length) ==0]  <- NA
+  res[lapply(res, length) == 0]  <- NA
   
   class(res) <- "gtrends"
   return(res)
 }
+